+++ app/src/main/java/com/hppk/sw/hppkcommuterbus/ui/MainActivity.kt
@@ -1,0 +1,22 @@
+package com.hppk.sw.hppkcommuterbus.ui
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import androidx.navigation.findNavController
+import androidx.navigation.ui.setupWithNavController
+import com.google.android.material.bottomnavigation.BottomNavigationView
+import com.hppk.sw.hppkcommuterbus.R
+
+
+class MainActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+
+        findViewById<BottomNavigationView>(R.id.bottomNavView).setupWithNavController(
+            findNavController(R.id.navHostFragment)
+        )
+    }
+
+}
