+++ app/src/main/java/com/hppk/sw/hppkcommuterbus/ui/busline/BusLinesFragment.kt
@@ -1,0 +1,22 @@
+package com.hppk.sw.hppkcommuterbus.ui.busline
+
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.hppk.sw.hppkcommuterbus.R
+
+class BusLinesFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_bus_lines, container, false)
+    }
+
+
+}
