+++ app/src/main/java/com/hppk/sw/hppkcommuterbus/ui/splash/SplashActivity.kt
@@ -2,13 +2,19 @@ package com.hppk.sw.hppkcommuterbus.ui.splash
 
 import android.annotation.SuppressLint
 import android.content.Intent
+import android.content.pm.PackageInfo
+import android.content.pm.PackageManager
 import android.os.Bundle
+import android.util.Base64
+import android.util.Log
 import android.view.Window
 import androidx.appcompat.app.AppCompatActivity
-import com.hppk.sw.hppkcommuterbus.MainActivity
 import com.hppk.sw.hppkcommuterbus.R
+import com.hppk.sw.hppkcommuterbus.ui.MainActivity
 import io.reactivex.Completable
 import io.reactivex.android.schedulers.AndroidSchedulers
+import java.security.MessageDigest
+import java.security.NoSuchAlgorithmException
 import java.util.concurrent.TimeUnit
 
 class SplashActivity : AppCompatActivity() {
@@ -22,6 +28,7 @@ class SplashActivity : AppCompatActivity() {
         Completable.timer(1, TimeUnit.SECONDS, AndroidSchedulers.mainThread())
             .subscribe(this::nextMoveActivity)
 
+        printHashKey() // TODO: for debug
     }
 
     private fun nextMoveActivity() {
@@ -30,4 +37,23 @@ class SplashActivity : AppCompatActivity() {
         startActivity(intent)
         finish()
     }
+
+    private fun printHashKey() {
+        try {
+            val info: PackageInfo = packageManager.getPackageInfo(
+                "com.hppk.sw.hppkcommuterbus",
+                PackageManager.GET_SIGNATURES
+            )
+            for (signature in info.signatures) {
+                val md = MessageDigest.getInstance("SHA")
+                md.update(signature.toByteArray())
+                Log.d("KeyHash:", Base64.encodeToString(md.digest(), Base64.DEFAULT))
+            }
+        } catch (e: PackageManager.NameNotFoundException) {
+            e.printStackTrace()
+        } catch (e: NoSuchAlgorithmException) {
+            e.printStackTrace()
+        }
+    }
+
 }
