+++ app/src/main/res/menu/menu_bottom_nav.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/mapFragment"
+        android:icon="@drawable/ic_map"
+        android:menuCategory="secondary"
+        android:title="@string/map"/>
+
+    <item
+        android:id="@+id/busLinesFragment"
+        android:icon="@drawable/ic_bus"
+        android:menuCategory="secondary"
+        android:title="@string/bus_lines"/>
+</menu>
