+++ app/src/main/res/navigation/nav_graph.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/nav_graph"
+    app:startDestination="@id/mapFragment">
+
+    <fragment
+        android:id="@+id/mapFragment"
+        android:name="com.hppk.sw.hppkcommuterbus.ui.map.MapFragment"
+        android:label="fragment_map"
+        tools:layout="@layout/fragment_map" />
+    <fragment
+        android:id="@+id/busLinesFragment"
+        android:name="com.hppk.sw.hppkcommuterbus.ui.busline.BusLinesFragment"
+        android:label="fragment_bus_lines"
+        tools:layout="@layout/fragment_bus_lines" />
+</navigation>
