+++ build.gradle
@@ -12,6 +12,8 @@ buildscript {
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
+
+        classpath "androidx.navigation:navigation-safe-args-gradle-plugin:2.1.0"
     }
 }
